:package: moving files
moving files to their respective folders
