+++ src/index.js
@@ -3,12 +3,12 @@ import './scss/app.scss';
 import './css/main.css';
 
 // import Task from './task';
-import renderNavbar from './navbar';
-import renderButton from './newTaskButton';
-import renderModal from './modal';
-import initializeLocalStorage from './initLocalStorage';
-import renderTasks from './renderTasks';
-import renderTitle from './renderTitle';
+import renderNavbar from './render/navbar';
+import renderButton from './render/newTaskButton';
+import renderModal from './render/modal';
+import initializeLocalStorage from './handlers/initLocalStorage';
+import renderTasks from './render/renderTasks';
+import renderTitle from './render/renderTitle';
 
 initializeLocalStorage();
 renderModal();
